pyproject.toml
[project]
name = "a-platform-for-publishing-paid-content"
version = "0.1.0"
description = ""
authors = [
    {name = "our ame",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "stripe (>=14.4.0,<15.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "coverage (>=7.13.4,<8.0.0)",
    "django-debug-toolbar (>=6.2.0,<7.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = false
warn_redundant_casts = false
warn_unused_configs = false
follow_imports = "skip"
exclude = [
    "migrations",
    "tests",
    "venv",
    ".venv",
    "__pycache__"
]

[[tool.mypy.overrides]]
module = [
    "django.*",
    "rest_framework.*",
    "stripe.*",
    "psycopg2.*",
    "dotenv.*",
    "jwt.*",
    "pytest.*",
]
ignore_missing_imports = true

[tool.black]
line-length = 119
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.nox
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/'''

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
